Fix more tests that launch ClusterStartup
